Add `policy` and `dynamic-config` sub-commands to `aperturectl apply` command
