+++ docs/content/reference/aperturectl/apply/apply.md
@@ -14,10 +14,6 @@ Apply Aperture Policy to the cluster
 
 Use this command to apply the Aperture Policy to the cluster.
 
-```
-aperturectl apply [flags]
-```
-
 ### Examples
 
 ```
@@ -29,13 +25,13 @@ aperturectl apply --dir=policy-dir
 ### Options
 
 ```
-      --dir string                   Path to directory containing Aperture Policy files
-      --dynamic-config-file string   Path to the dynamic config file
-      --file string                  Path to Aperture Policy file
-  -h, --help                         help for apply
-      --kube-config string           Path to the Kubernetes cluster config. Defaults to '~/.kube/config'
+      --dir string           Path to directory containing Aperture Policy files
+      --file string          Path to Aperture Policy file
+  -h, --help                 help for apply
+      --kube-config string   Path to the Kubernetes cluster config. Defaults to '~/.kube/config'
 ```
 
 ### SEE ALSO
 
 - [aperturectl](/reference/aperturectl/aperturectl.md) - aperturectl - CLI tool to interact with Aperture
+- [aperturectl apply policy](/reference/aperturectl/apply/policy/policy.md) - Apply Aperture Policy to the cluster
