+++ docs/content/reference/aperturectl/apply/policy/policy.md
@@ -1,0 +1,37 @@
+---
+sidebar_label: Policy
+hide_title: true
+keywords:
+  - aperturectl
+  - aperturectl_apply_policy
+---
+
+## aperturectl apply policy
+
+Apply Aperture Policy to the cluster
+
+### Synopsis
+
+Use this command to apply the Aperture Policy to the cluster.
+
+```
+aperturectl apply policy [flags]
+```
+
+### Examples
+
+```
+aperturectl apply policy --file=policy.yaml
+
+aperturectl apply policy --dir=policy-dir
+```
+
+### Options
+
+```
+  -h, --help   help for policy
+```
+
+### SEE ALSO
+
+- [aperturectl apply](/reference/aperturectl/apply/apply.md) - Apply Aperture Policy to the cluster
